refactor: Make logic more section oriented
First approach was very focused on landing containers.
Working with sections is much clearer.

Removed all helper functions and switched
to nested functions completely to keep the
global namespace as clean as possible.
Still trying to stick to the general layout of the exercise template.

Clarified some pieces of documentation.